# PERF: removed coercion to int64 for arrays of ints in Categorical.from_codes