--- a/pandas/core/arrays/categorical.py
+++ b/pandas/core/arrays/categorical.py
@@ -578,7 +578,11 @@ class Categorical(ExtensionArray, PandasObject):
             unordered.
         """
         try:
-            codes = np.asarray(codes, np.int64)
+            if (type(codes) == np.ndarray
+                    and np.issubdtype(codes.dtype, np.integer)):
+                codes = np.asarray(codes)
+            else:
+                codes = np.asarray(codes, np.int64)
         except (ValueError, TypeError):
             raise ValueError(
                 "codes need to be convertible to an arrays of integers")