@TomAugspurger 2018-05-11
Could you also add a release note (performance improvement)?