@fivemok 2018-05-12
Updated ASV:
```
       before           after         ratio
     [3d03fdb2]       [d83abca4]
       9.27±0.6ms       9.13±0.3ms     0.98  categoricals.Concat.time_concat
       5.80±0.1ms      6.22±0.03ms     1.07  categoricals.Concat.time_union
       32.7±0.6ms       32.8±0.4ms     1.00  categoricals.Constructor.time_all_nan
      1.26±0.03ms      1.24±0.01ms     0.99  categoricals.Constructor.time_datetimes
         1.28±0ms      1.28±0.01ms     1.00  categoricals.Constructor.time_datetimes_with_nat
          364±8μs          348±5μs     0.96  categoricals.Constructor.time_fastpath
      19.9±0.08ms      20.0±0.06ms     1.00  categoricals.Constructor.time_regular
        181±0.5ms          182±2ms     1.00  categoricals.Constructor.time_with_nan
           10.2ms           10.3ms     1.01  categoricals.Isin.time_isin_categorical('int64')
       10.8±0.1ms      10.5±0.09ms     0.97  categoricals.Isin.time_isin_categorical('object')
      8.89±0.08ms       8.94±0.1ms     1.01  categoricals.Rank.time_rank_int
       9.33±0.2ms      9.44±0.05ms     1.01  categoricals.Rank.time_rank_int_cat
      8.95±0.08ms      8.95±0.05ms     1.00  categoricals.Rank.time_rank_int_cat_ordered
          135±2ms          140±2ms     1.03  categoricals.Rank.time_rank_string
      11.0±0.06ms       11.1±0.2ms     1.01  categoricals.Rank.time_rank_string_cat
       9.14±0.2ms       9.11±0.1ms     1.00  categoricals.Rank.time_rank_string_cat_ordered
          590±7μs         599±20μs     1.01  categoricals.Repr.time_rendering
       27.7±0.5ms       29.1±0.2ms     1.05  categoricals.SetCategories.time_set_categories
       30.0±0.4ms       29.6±0.4ms     0.99  categoricals.ValueCounts.time_value_counts(False)
       29.6±0.4ms       29.6±0.2ms     1.00  categoricals.ValueCounts.time_value_counts(True)```