@TomAugspurger 2018-05-12
I think @jbreack was requesting a *new* ASV that measures `Categorical.from_codes` with an `int8`  type array.

It can go in `asv_bench/benchmarks/categoricals.py` under `Constructor`.